refactor(db): split pg.py into per-entity submodules behind a shim
pg.py grew to 4910 lines and ate every CRUD function the project has.
This commit splits it into a db/ package by entity, with no semantic
changes — every function body is byte-identical to its old position.

Layout:
  _pool.py    connection pools + lifecycle (init/close/_init_test_database)
  schema.py   pure DDL (CREATE TABLE / RLS policies)
  tenant.py   Tenant CRUD
  user.py     User CRUD + OIDC token vault
  coworker.py Coworker CRUD + user-agent assignments
  chat.py     ChannelBinding / Conversation / Session / Message
  task.py     Scheduled tasks + run logs
  legacy.py   Migration helpers (RegisteredGroup, sessions_legacy)
  approval.py Approval policies / requests / audit
  safety.py   Safety rules / decisions / audit
  skill.py    Skills + skill files

pg.py is reduced to a re-export shim so the 79 existing
``from rolemesh.db.pg import X`` call sites keep working unchanged.
A follow-up PR migrates them to ``from rolemesh.db import X`` and
deletes the shim.

Verified zero-semantic-change:
  * each submodule body diff'd byte-for-byte against its slice of the
    pre-refactor pg.py
  * mypy errors drop from 106 to 99 (the seven that were duplicated
    between pg.py and its physical copies are now counted once)
  * tests/db (77 tests, RLS + cross-tenant isolation + connection
    wrappers + skills + token columns + function classification) all
    green
  * ruff strict (TCH/F/I/UP/B/SIM/RUF) passes on the new package
